% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_sql.R
\docType{package}
\name{knitsql}
\alias{knitsql}
\alias{knitsql-package}
\title{DBI/SQL knitr engine for R}
\description{
Call \code{library(knitsql)} to automatically register a \code{sql} knitr
engine. This allows you to create R Markdown code chunks that contain SQL
queries.
}
\details{
TODO: Improve these docs

Specify the connection either in R Markdown metadata, or by creating a DBI
connection object in a prior R code chunk and pass it as the \code{conn}
object.

Use the knitr chunk name to assign the results to a variable.

You can bind R variables to parameters in the SQL query by using
\code{?varname}.
}

